# use pool for less contention