--- a/uuid_test.go
+++ b/uuid_test.go
@@ -7,6 +7,8 @@ import (
 )
 
 func TestUUID(t *testing.T) {
+	// TODO fix this test now that we can return non-deterministic
+	// counters.
 	var buf [24]byte
 	for i := range buf {
 		buf[i] = byte(i) + 1
@@ -32,13 +34,29 @@ func TestUUID(t *testing.T) {
 
 func TestUniqueness(t *testing.T) {
 	g := MustNewGenerator()
+	mc := make(chan map[[24]byte]int)
+	const nproc = 4
+	for i := 0; i < nproc; i++ {
+		go func() {
+			m := make(map[[24]byte]int)
+			for i := 0; i < step*10; i++ {
+				uuid := g.Next()
+				if old, ok := m[uuid]; ok {
+					t.Errorf("non-unique uuid seq at %d, other %d", i, old)
+				}
+				m[uuid] = i
+			}
+			mc <- m
+		}()
+	}
 	m := make(map[[24]byte]int)
-	for i := 0; i < 65537; i++ {
-		uuid := g.Next()
-		if old, ok := m[uuid]; ok {
-			t.Fatalf("non-unique uuid seq at %d, other %d", i, old)
+	for i := 0; i < nproc; i++ {
+		for uuid, iter := range <-mc {
+			if old, ok := m[uuid]; ok {
+				t.Errorf("non-unique uuid seq at %d, other %d", i, old)
+			}
+			m[uuid] = iter
 		}
-		m[uuid] = i
 	}
 }
 
@@ -47,4 +65,13 @@ func BenchmarkNext(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		g.Next()
 	}
+}
+
+func BenchmarkContended(b *testing.B) {
+	g := MustNewGenerator()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			g.Next()
+		}
+	})
 }