This makes it somewhat slower in the single-threaded case but much faster when contended.
    
    name         old time/op  new time/op  delta
    Contended    18.3ns ± 1%  21.9ns ± 3%  +20.04%  (p=0.008 n=5+5)
    Contended-2  33.5ns ± 1%  10.6ns ± 1%  -68.41%  (p=0.016 n=5+4)