@rogpeppe 2019-06-13
Closing as it's probably more complexity than we want to take on board, and makes the uncontended case slower.